refactor: fixture 클래스 분리
